refactor: deduplicate shared code across commands
- Extract BaseExecutionConfig type and prepareExecution() into
  src/commands/execution.ts, eliminating ~170 lines of identical
  setup code between capture and trials (schema loading, prompt
  loading, path resolution, session manager, output init, workspace)

- Add executePrompts() wrapper for common worker pool invocation
  pattern (progress callback, error logging, completion)

- Add parseConcurrency() CLI helper shared by capture and trials

- Add loadGraderOrExit() in grader-loader.ts to replace 4 identical
  try/catch blocks across capture, trials, calibrate, validate-refs

- Remove 4 local copies of resolvePath() (balance, validate-refs,
  grader-loader, schemas-cli) — now imported from core

- CaptureConfig = BaseExecutionConfig (type alias)
- TrialsConfig = BaseExecutionConfig & { k: number }
